// client/src/app/api/models/championship-row-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChampionshipRowDto {
  constructorsChampId?: string;
  constructorsChampName?: string | null;
  driversChampId?: string;
  driversChampName?: string | null;
  season?: string | null;
  seriesName?: string | null;
  status?: string | null;
}