// client/src/app/api/models/participation-add-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { DriverParticipationRowDto } from './driver-participation-row-dto';
export interface ParticipationAddDto {
  constructorIds?: Array<string> | null;
  constructorsChampId?: string;
  drivers?: Array<DriverParticipationRowDto> | null;
  driversChampId?: string;
}